Rename asserts package to pytest
